+++ src/Slices/Diagnose/Core/Domain.ts
@@ -1,5 +1,5 @@
 import { ParsedNumber } from "@/Core/Language";
-import { CompileError } from "./CompileError";
+import { CompileError } from "../../../Core/Domain/CompileError";
 
 export interface RawDiagnostics {
   rejections: RawRejection[];
