+++ src/Slices/Diagnose/Data/StateHelper.ts
