+++ src/Slices/Diagnose/UI/Diagnose.tsx
