+++ src/Slices/Diagnose/UI/Page.tsx
