+++ src/Slices/Diagnose/index.ts
@@ -1,0 +1,1 @@
+export * as Diagnose from "./Core/Route";
