+++ src/UI/Pages/PrimaryPageManager.tsx
@@ -5,6 +5,7 @@ import { AgentsPage } from "@S/Agents/UI";
 import { CompileDetailsPage } from "@S/CompileDetails/UI";
 import { CompileReportsPage } from "@S/CompileReports/UI";
 import { ComplianceCheckPage } from "@S/ComplianceCheck/UI";
+import { DiagnosePage } from "@S/Diagnose/UI";
 import { EventsPage } from "@S/Events/UI";
 import { FactsPage } from "@S/Facts/UI";
 import { NotificationCenterPage } from "@S/Notification/UI/Center";
@@ -17,7 +18,6 @@ import { DesiredStatePage } from "./DesiredState";
 import { DesiredStateComparePage } from "./DesiredStateCompare";
 import { DesiredStateDetailsPage } from "./DesiredStateDetails";
 import { DesiredStateResourceDetailsPage } from "./DesiredStateResourceDetails";
-import { DiagnosePage } from "./Diagnose";
 import { EditInstancePage } from "./EditInstance";
 import { HomePage } from "./Home";
 import { ParametersPage } from "./Parameters";
