+++ src/UI/Routing/PrimaryRouteManager.ts
@@ -18,6 +18,7 @@ import { Agents } from "@S/Agents";
 import { CompileDetails } from "@S/CompileDetails";
 import { CompileReports } from "@S/CompileReports";
 import { ComplianceCheck } from "@S/ComplianceCheck";
+import { Diagnose } from "@S/Diagnose";
 import { Events } from "@S/Events";
 import { Facts } from "@S/Facts";
 import { Notification } from "@S/Notification";
@@ -49,7 +50,7 @@ export class PrimaryRouteManager implements RouteManager {
       CreateInstance: CreateInstance(this.baseUrl),
       EditInstance: EditInstance(this.baseUrl),
       History: ServiceInstanceHistory.route(this.baseUrl),
-      Diagnose: Diagnose(this.baseUrl),
+      Diagnose: Diagnose.route(this.baseUrl),
       Events: Events.route(this.baseUrl),
 
       /**
@@ -212,14 +213,6 @@ const EditInstance = (base: string): Route<"EditInstance"> => ({
   parent: "Inventory",
   path: `${base}${paths.EditInstance}`,
   generateLabel: () => "Edit Instance",
-  environmentRole: "Required",
-});
-
-const Diagnose = (base: string): Route<"Diagnose"> => ({
-  kind: "Diagnose",
-  parent: "Inventory",
-  path: `${base}${paths.Diagnose}`,
-  generateLabel: () => "Diagnose Service Instance",
   environmentRole: "Required",
 });
